backend/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
numpy
pillow
fastapi-cors
